rename analytics to telemetry
