+++ liquibase-standard/src/main/java/liquibase/Scope.java
@@ -1,6 +1,6 @@
 package liquibase;
 
-import liquibase.analytics.Event;
+import liquibase.telemetry.Event;
 import liquibase.configuration.LiquibaseConfiguration;
 import liquibase.database.Database;
 import liquibase.database.DatabaseConnection;
