--- liquibase-standard/src/main/java/liquibase/analytics/AnalyticsOutputDestination.java
@@ -1,5 +1,0 @@
-package liquibase.analytics;
-
-public enum AnalyticsOutputDestination {
-    SEGMENT, CSV
-}
