+++ liquibase-standard/src/main/java/liquibase/telemetry/SegmentBatch.java
@@ -1,4 +1,4 @@
-package liquibase.analytics;
+package liquibase.telemetry;
 
 import lombok.Data;
