+++ liquibase-standard/src/main/java/liquibase/telemetry/SegmentTrackEvent.java
@@ -1,4 +1,4 @@
-package liquibase.analytics;
+package liquibase.telemetry;
 
 import liquibase.Scope;
 import liquibase.license.LicenseService;
