+++ liquibase-standard/src/main/java/liquibase/telemetry/TelemetryOutputDestination.java
@@ -1,0 +1,5 @@
+package liquibase.telemetry;
+
+public enum TelemetryOutputDestination {
+    SEGMENT, CSV
+}
