+++ liquibase-standard/src/main/resources/META-INF/services/liquibase.configuration.AutoloadedConfigurations
@@ -3,4 +3,4 @@ liquibase.integration.commandline.LiquibaseCommandLineConfiguration
 liquibase.logging.core.DefaultLoggerConfiguration
 liquibase.sql.SqlConfiguration
 liquibase.parser.ChangeLogParserConfiguration
-liquibase.analytics.TelemetryConfiguration
+liquibase.telemetry.TelemetryConfiguration
